requirements.txt
astroid==2.3.2
autopep8==1.4.4
Babel==2.7.0
colorama==0.4.1
Django==2.2.6
django-colorfield==0.1.15
django-colorful==1.3
django-crispy-forms==1.8.0
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.10.21
Pillow==6.2.1
pycodestyle==2.5.0
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
